IGNITE-14477: fixed for tests
