+++ modules/ducktests/src/main/java/org/apache/ignite/internal/ducktest/tests/rebalance/DataModel.java
@@ -1,0 +1,51 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.ignite.internal.ducktest.tests.rebalance;
+
+import java.util.concurrent.ThreadLocalRandom;
+
+/**
+ * Data model class, which instances used as cache entry values.
+ */
+public class DataModel {
+    /** Cached payload. */
+    private static byte[] cachedPayload;
+
+    /** Payload. */
+    private final byte[] payload;
+
+    /**
+     * @param entrySize Entry size.
+     */
+    DataModel(int entrySize) {
+        payload = getPayload(entrySize);
+    }
+
+    /**
+     * @param payloadSize Payload size.
+     */
+    private static byte[] getPayload(int payloadSize) {
+        if (cachedPayload == null || cachedPayload.length != payloadSize) {
+            cachedPayload = new byte[payloadSize];
+
+            ThreadLocalRandom.current().nextBytes(cachedPayload);
+        }
+
+        return cachedPayload;
+    }
+}
